requirements/requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
anthropic>=0.8.0
langchain-openai>=0.0.5
chromadb>=0.4.22
loguru>=0.7.2
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
